requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
transformers>=4.36.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
einops>=0.6.1
timm>=0.9.2
nltk>=3.8
editdistance>=0.6.2
pycocotools>=2.0.6
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tensorboard>=2.13.0

# Additional dependencies for baseline models
accelerate>=0.20.0
bitsandbytes>=0.41.0
flash-attn>=2.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0
datasets>=2.14.0
peft>=0.5.0

# Vision-Language model dependencies  
qwen-vl-utils>=0.0.1
llava>=1.2.0
open-clip-torch>=2.20.0

# Gemma-3 VLM (PaliGemma) dependencies
google-generativeai>=0.3.0
paligemma>=0.1.0
siglip>=0.1.0

# Additional utilities
pandas>=2.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0
